+++ docs/source/conf.py
@@ -293,7 +293,7 @@ texinfo_documents = [
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'https://docs.python.org/2/': None,
+    'https://docs.python.org/3/': None,
     'boto': ('http://boto.readthedocs.org/en/latest/', None),
     'boto3': ('http://boto3.readthedocs.org/en/latest/', None)
 }