+++ CHANGES.rst
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+Unreleased Changes
+------------------
+
+* `PR #276 <https://github.com/jantman/awslimitchecker/pull/276>`_ /
+  `Issue #275 <https://github.com/jantman/awslimitchecker/issues/275>`_ - Add new
+  ``--skip-service`` CLI option and :py:meth:`~.Checker.remove_services` to allow
+  skipping of one or more specific services during runs. (Thanks to `tamsky <https://github.com/tamsky>`_)
+  for this contribution.)
+* `PR #274 <https://github.com/jantman/awslimitchecker/pull/274>`_ /
+  `Issue #273 <https://github.com/jantman/awslimitchecker/issues/273>`_ - Add support
+  for new ``i3`` EC2 Instance types.  (Thanks to `tamsky <https://github.com/tamsky>`_)
+  for this contribution.)
+
 0.8.0 (2017-03-11)
 ------------------